backend/requirements.txt
aiodns==3.0.0
aiofile==3.8.5
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
caio==0.9.12
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
pytz==2023.3
requests==2.30.0
six==1.16.0
stream-chat==4.8.0
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
